separate components by feature
