--- react-ui/src/components/NotFound/index.js
@@ -1,18 +1,0 @@
-import Background from "../layouts/Background";
-
-export default function NotFound() {
-    return (
-        <Background>
-            <div className="row d-flex justify-content-center">
-                <section className="mr-auto ml-auto text-center">
-                    <img  
-                        src="https://t3.ftcdn.net/jpg/01/67/37/06/240_F_167370626_zEV8195DIsvDnCJHIkBbBHgIotS4vPKl.jpg" 
-                        alt="spilled coffee" 
-                    />
-                    <h2>That was Strange...</h2>
-                    <p>The page you were looking for could not be found.</p>
-                </section>
-            </div>
-        </Background>
-    );
-}
