--- react-ui/src/components/ReadingLog/ReadingLogAdd/index.js
@@ -1,8 +1,0 @@
-import ReadingLogSplitLayout from "../ReadingLogSplitLayout";
-import ReadingLogAddForm from "./ReadingLogAddForm";
-
-export default function ReadingLogAdd() {    
-    return (
-        <ReadingLogSplitLayout component={<ReadingLogAddForm />} pageTitle="Add Reading Log" />
-    );
-}
