+++ react-ui/src/components/layouts/PageLayout.js
@@ -1,0 +1,11 @@
+import Title from "../Title";
+import Background from "./Background";
+
+export default function PageLayout({ pageTitle, children }) {
+    return (
+        <Background>
+            <Title text={pageTitle} />
+            {children}
+        </Background>
+    )
+}
