+++ react-ui/src/features/Bookshelf/hoc/withBooks.js
