+++ react-ui/src/features/ReadingLog/components/ReadingLogAdd/index.js
@@ -1,0 +1,11 @@
+import ReadingLogSplitLayout from "../../layouts/ReadingLogSplitLayout";
+import ReadingLogAddForm from "./ReadingLogAddForm";
+
+export default function ReadingLogAdd() {    
+    return (
+        <ReadingLogSplitLayout 
+            component={<ReadingLogAddForm />} 
+            pageTitle="Add Reading Log" 
+        />
+    );
+}
